add support for aws cognito jwt
cognito tokens do not have a typ field in the header so we have to skip the test.

the issuer url contain a mandatory path representing the pool id. We have to account for it when we check the issuer and also when we crate a new url to fetch the keys.

an example of issuer url:
https://cognito-idp.us-east-1.amazonaws.com/us-east-1_x4phO7R6v